+++ randovania/games/cave_story/exporter/game_exporter.py
@@ -13,10 +13,13 @@ from randovania.lib import json_lib, status_update_lib
 if TYPE_CHECKING:
     from pathlib import Path
 
+    from caver_patcher import CSPlatform
+
 
 @dataclasses.dataclass(frozen=True)
 class CSGameExportParams(GameExportParams):
     output_path: Path
+    platform: CSPlatform
 
 
 class CSGameExporter(GameExporter):
@@ -53,7 +56,9 @@ class CSGameExporter(GameExporter):
         new_patch = copy.copy(patch_data)
         if new_patch["mychar"] is not None:
             new_patch["mychar"] = str(RandovaniaGame.CAVE_STORY.data_path.joinpath(patch_data["mychar"]))
+
+        new_patch["platform"] = export_params.platform.value
         try:
-            caver_patcher.patch_files(new_patch, export_params.output_path, progress_update)
+            caver_patcher.patch_files(new_patch, export_params.output_path, export_params.platform, progress_update)
         finally:
             json_lib.write_path(export_params.output_path.joinpath("data", "patcher_data.json"), patch_data)