+++ randovania/games/cave_story/exporter/options.py
@@ -3,6 +3,8 @@ from __future__ import annotations
 import dataclasses
 from pathlib import Path
 
+from caver.patcher import CSPlatform
+
 from randovania.games.game import RandovaniaGame
 from randovania.interface_common.options import PerGameOptions, decode_if_not_none
 
@@ -10,12 +12,14 @@ from randovania.interface_common.options import PerGameOptions, decode_if_not_no
 @dataclasses.dataclass(frozen=True)
 class CSPerGameOptions(PerGameOptions):
     output_directory: Path | None = None
+    platform: CSPlatform = CSPlatform.FREEWARE
 
     @property
     def as_json(self):
         return {
             **super().as_json,
             "output_directory": str(self.output_directory) if self.output_directory is not None else None,
+            "platform": self.platform.value,
         }
 
     @classmethod
@@ -25,4 +29,5 @@ class CSPerGameOptions(PerGameOptions):
         return cls(
             cosmetic_patches=cosmetic_patches,
             output_directory=decode_if_not_none(value["output_directory"], Path),
+            platform=CSPlatform(value["platform"]),
         )